\name{monitoring}
\alias{monitoring}
\docType{package}
\title{
Collection of functions for monitoring sequences over time: Shewhart, EWMA, and Cusum routines
}
\description{
Collection of functions for monitoring sequences over time: Shewhart, EWMA, and Cusum routines
}
\details{
\tabular{ll}{
Package: \tab monitoring\cr
Type: \tab Package\cr
Version: \tab 2013.04.22\cr
Date: \tab 2013-04-22\cr
License: \tab Unknown \cr
}
}
\author{
Maintainer: Landon Sego <Landon.Sego@pnl.gov>
}
% \references{
% ~~ Literature or other references for background information ~~
% }
\keyword{package}
% ~~ Optionally other keywords from doc/KEYWORDS, one per line
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
